Took out the join table cds_service_endpoint_cds_hooks.  Added column cds_service_endpoint_id to cds_hook table.  Took out cds_service_endpoint object from LaunchScenario.  Added code to save images for CDS-Hook.

+++ src/main/java/org/hspconsortium/sandboxmanagerapi/controllers/CdsHookController.java
@@ -2,8 +2,11 @@ package org.hspconsortium.sandboxmanagerapi.controllers;
 
 import com.amazonaws.services.cloudwatch.model.ResourceNotFoundException;
 import org.hspconsortium.sandboxmanagerapi.model.CdsHook;
+import org.hspconsortium.sandboxmanagerapi.model.CdsServiceEndpoint;
 import org.hspconsortium.sandboxmanagerapi.model.Image;
+import org.hspconsortium.sandboxmanagerapi.services.AuthorizationService;
 import org.hspconsortium.sandboxmanagerapi.services.CdsHookService;
+import org.hspconsortium.sandboxmanagerapi.services.CdsServiceEndpointService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.MediaType;
@@ -18,28 +21,23 @@ import java.io.IOException;
 
 import static org.springframework.http.MediaType.*;
 
-// TODO: add endpoint to save images and cdsHook
 @RestController
 @RequestMapping({"/cds-hook"})
 public class CdsHookController {
     private static Logger LOGGER = LoggerFactory.getLogger(AppController.class.getName());
 
     private final CdsHookService cdsHookService;
+    private final AuthorizationService authorizationService;
+    private final CdsServiceEndpointService cdsServiceEndpointService;
 
     @Inject
-    public CdsHookController(final CdsHookService cdsHookService) {
+    public CdsHookController(final CdsHookService cdsHookService,
+                             final AuthorizationService authorizationService,
+                             final CdsServiceEndpointService cdsServiceEndpointService) {
         this.cdsHookService = cdsHookService;
+        this.authorizationService = authorizationService;
+        this.cdsServiceEndpointService = cdsServiceEndpointService;
     }
-
-    @PostMapping
-    @Transactional
-    @ResponseBody
-    public CdsHook saveImage(final HttpServletRequest request,
-                                            @RequestBody CdsHook cdsServiceEndpoint) {
-        return null;
-    }
-
-    //************************  Image Saving from appController ***************************
 
     @GetMapping(value = "/{id}/image", produces ={IMAGE_GIF_VALUE, IMAGE_PNG_VALUE, IMAGE_JPEG_VALUE, "image/jpg"})
     @ResponseBody
@@ -60,10 +58,11 @@ public class CdsHookController {
     @Transactional
     public @ResponseBody void putFullImage(final HttpServletRequest request, @PathVariable Integer id, @RequestParam("file") MultipartFile file) {
         CdsHook cdsHook = cdsHookService.getById(id);
+        CdsServiceEndpoint cdsServiceEndpoint = cdsServiceEndpointService.getById(cdsHook.getCdsServiceEndpointId());
         if (cdsHook == null) {
             throw new ResourceNotFoundException("CDS-Hook does not exist. Cannot upload image.");
         }
-//        authorizationService.checkSandboxUserModifyAuthorization(request, cdsHook.getSandbox(), cdsHook);
+        authorizationService.checkSandboxUserModifyAuthorization(request, cdsServiceEndpoint.getSandbox(), cdsServiceEndpoint);
         cdsHook.setLogoUri(request.getRequestURL().toString());
         cdsHookService.save(cdsHook);
         try {
@@ -82,12 +81,11 @@ public class CdsHookController {
     @Transactional
     public CdsHook deleteFullImage(final HttpServletRequest request, @PathVariable Integer id) {
         CdsHook cdsHook = cdsHookService.getById(id);
+        CdsServiceEndpoint cdsServiceEndpoint = cdsServiceEndpointService.getById(cdsHook.getCdsServiceEndpointId());
         if (cdsHook == null) {
             throw new ResourceNotFoundException("CDS-Hook does not exist. Cannot delete image.");
         }
-//        authorizationService.checkSandboxUserModifyAuthorization(request, app.getSandbox(), app);
+        authorizationService.checkSandboxUserModifyAuthorization(request, cdsServiceEndpoint.getSandbox(), cdsServiceEndpoint);
         return cdsHookService.deleteCdsHookImage(cdsHook);
     }
-
-    //***********************************************************************************************
 }
